ServiApp/requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.6
cryptography==37.0.4
cycler==0.11.0
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.0
django-redis==5.2.0
django-rq==2.5.1
djangorestframework==3.13.1
firebase-admin==2.13.0
fonttools==4.38.0
gcloud==0.18.3
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
joblib==1.2.0
jwcrypto==1.3.1
kiwisolver==1.4.4
matplotlib==3.6.2
mlxtend==0.21.0
msgpack==1.0.4
mssql-django==1.1.3
numpy==1.23.4
oauth2client==4.1.3
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
proto-plus==1.22.0
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydot==1.4.2
pydotplus==2.0.2
pyodbc==4.0.34
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
requests-toolbelt==0.9.1
rq==1.11.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
wrapt==1.14.1
